Revamped to add full width layout.
The form elements are now all correctly namespaced to the search instance, as well as the scopes (when enabled). Added the method `makeFullWidth()` to switch the search bar to a full width bar below the subnavigation.

Also added more public methods to access information, including the `getPersistVars()` method, which allows fetching all variables needed to restore the search via filter settings, or a data grid for example.